requirements.txt
paramiko==2.10.4
tornado==5.1.1; python_version < '3.5'
tornado==6.1.0; python_version >= '3.5'
gql[all]==3.4.0
python-dotenv==0.20.0
